refactor(interviews/shortest_common_prefixes)
Split into v1 & v2
