+++ .pylintrc
@@ -5,6 +5,7 @@ disable=
     C0116, # missing-function-docstring
     W0105, # pointless-string-statement
     W0511, # fixme
+    R0801, # similar lines
 
 [FORMAT]
 max-line-length=150
