+++ src/problems/interviews/shortest_common_prefixes/v1.py
@@ -1,8 +1,5 @@
 import string
-from typing import Any, Iterator, Optional
-
-import more_itertools
-from hamcrest import assert_that, equal_to
+from typing import Any
 
 from src.problems.config import DEBUG
 
@@ -42,17 +39,6 @@ def pluck(data: dict[str, Any | str]) -> list[str]:
     return results
 
 
-def pluck2(data: dict[str, Any | str]) -> Iterator[Any]:
-    def recurse(dictionary: dict[str, Any | str]) -> list[Any | str]:
-        return [
-            recurse(value) if isinstance(value, dict) else key  # type: ignore
-            for key, value in dictionary.items()
-        ]
-
-    results = recurse(data)
-    return more_itertools.collapse(results)
-
-
 def get_prefixes(words: list[str]):
     data = categorize(words)
     if DEBUG:
@@ -70,21 +56,4 @@ def get_prefixes(words: list[str]):
     # }
 
     # TODO benchmark pluck VS pluck2 performance
-    return list(pluck2(data))
-
-
-def test(words: list[str], expected: Optional[list[str]] = None):
-    actual = get_prefixes(words)
-    assert_that(expected, equal_to(actual))
-    print("✅", end="")
-
-
-def run():
-    test([], expected=[])
-    test(["dog", "zebra", "bananas"], expected=["b", "d", "z"])
-    test(["dog", "zebra", "duck", "bananas"], expected=["b", "do", "du", "z"])
-    test(["dog", "zebra", "duck", "dove", "bananas"], expected=["b", "dog", "dov", "du", "z"])
-
-
-if __name__ == "__main__":
-    run()
+    return list(pluck(data))
