+++ tests/step_defs/problems/interview/test__shortest_common_prefixes.py
@@ -1,7 +1,7 @@
 from hamcrest import assert_that, equal_to
 from pytest_bdd import given, parsers, scenarios, then, when
 
-from src.problems.interviews.shortest_common_prefixes import get_prefixes
+from src.problems.interviews.shortest_common_prefixes.v2 import get_prefixes
 
 scenarios("problems/interviews/Shortest Common Prefixes.feature")
